Add files via upload (#5)
Updated code, not much has changed and there is a lot of unfinished code. The player class is not done yet. It still has to give the user top three times in sorted order, get all times for a specific user, and check that name is valid and new players don't give usernames already in the file. We also have to create a way to present our game. Text based user interface is acceptable and using java swing to create a layout.